test(ui): split chat catalog pane coverage
