+++ ui/src/pages/chat/chat-pane-history.test.ts
@@ -1,18 +1,13 @@
 /* @vitest-environment jsdom */
 
 import { describe, expect, it, vi } from "vitest";
-import type {
-  SessionCatalogSession,
-  SessionCatalogTranscriptItem,
-} from "../../../../packages/gateway-protocol/src/index.js";
+import type { SessionCatalogTranscriptItem } from "../../../../packages/gateway-protocol/src/index.js";
 import { createDeferred } from "../../../../test/helpers/promise.js";
 import type { GatewayBrowserClient } from "../../api/gateway.ts";
 import type { ApplicationContext } from "../../app/context.ts";
-import { buildCatalogSessionKey, type CatalogSessionKey } from "../../lib/sessions/catalog-key.ts";
 import type { SessionCapability } from "../../lib/sessions/index.ts";
 import "./chat-pane.ts";
 import { loadChatHistory } from "./chat-history.ts";
-import { consumePaneSessionHandoff } from "./chat-pane-shared.ts";
 import type { ChatPageHost } from "./chat-state-host.ts";
 
 type TestChatPane = HTMLElement & {
@@ -21,12 +16,7 @@ type TestChatPane = HTMLElement & {
   state: ChatPageHost;
   connectedClient: GatewayBrowserClient | null;
   connectionGeneration: number;
-  continueCatalogSession: (key: CatalogSessionKey) => Promise<void>;
-  catalogLoadGeneration: number;
-  catalogSession: SessionCatalogSession | null;
-  paneId: string;
   sessionKey: string;
-  onPaneSessionChange?: (paneId: string, sessionKey: string) => void;
   catalogItemMessage: (item: SessionCatalogTranscriptItem) => Record<string, unknown> | null;
   handleTranscriptScroll: (event: Event) => void;
   historyAutoLoadBlocked: boolean;
@@ -121,34 +111,6 @@ function createTestChatPane(params: { client: GatewayBrowserClient; sessions: Se
   pane.connectedClient = params.client;
   pane.connectionGeneration = 4;
   return { pane, state, requestUpdate };
-}
-
-function createCatalogContinuationPane(request: ReturnType<typeof vi.fn>) {
-  const client = { request } as unknown as GatewayBrowserClient;
-  const sessions = {} as SessionCapability;
-  const { pane, requestUpdate, state } = createTestChatPane({ client, sessions });
-  const key = {
-    catalogId: "codex",
-    hostId: "gateway:local",
-    threadId: "thread-101",
-  } satisfies CatalogSessionKey;
-  const sourceSessionKey = buildCatalogSessionKey(key);
-  state.sessionKey = sourceSessionKey;
-  pane.sessionKey = sourceSessionKey;
-  state.chatMessage = "Continue the original catalog conversation";
-  state.handleChatDraftChange = vi.fn((draft: string) => {
-    state.chatMessage = draft;
-  });
-  state.handleSendChat = vi.fn(async () => undefined);
-  pane.catalogSession = {
-    threadId: key.threadId,
-    status: "idle",
-    archived: false,
-    canContinue: true,
-    canArchive: true,
-  };
-  pane.onPaneSessionChange = vi.fn();
-  return { client, key, pane, requestUpdate, sessions, sourceSessionKey, state };
 }
 
 function nativeHistoryMessage(seq: number, text = `message ${seq}`) {
@@ -697,278 +659,5 @@ describe("chat pane native history pagination", () => {
     await loadChatHistory(state);
 
     expect(state.chatHistoryPagination).toBe(pagination);
-  });
-});
-
-describe("chat pane catalog continuation lifecycle", () => {
-  it("hands a continued catalog draft to the retained destination pane", async () => {
-    const request = vi.fn().mockResolvedValue({ sessionKey: "agent:main:continued" });
-    const { key, pane, state } = createCatalogContinuationPane(request);
-    pane.catalogSession = { ...pane.catalogSession!, sourceHomeId: "source-home-a" };
-
-    await pane.continueCatalogSession(key);
-
-    expect(request).toHaveBeenCalledWith("sessions.catalog.continue", {
-      ...key,
-      agentId: "main",
-      sourceHomeId: "source-home-a",
-    });
-    expect(pane.onPaneSessionChange).toHaveBeenCalledWith("single", "agent:main:continued");
-    expect(consumePaneSessionHandoff(pane.context, pane.paneId, "agent:main:continued")).toEqual({
-      attachments: [],
-      draft: "Continue the original catalog conversation",
-      send: true,
-    });
-    expect(state.sessionKey).not.toBe("agent:main:continued");
-    expect(state.handleSendChat).not.toHaveBeenCalled();
-  });
-
-  it("carries staged attachments through the continuation handoff", async () => {
-    const request = vi.fn().mockResolvedValue({ sessionKey: "agent:main:continued-attachments" });
-    const { key, pane, state } = createCatalogContinuationPane(request);
-    state.chatAttachments = [
-      {
-        id: "att-1",
-        mimeType: "image/png",
-        fileName: "shot.png",
-        dataUrl: "data:image/png;base64,AAAA",
-      },
-    ];
-
-    await pane.continueCatalogSession(key);
-
-    const handoff = consumePaneSessionHandoff(
-      pane.context,
-      pane.paneId,
-      "agent:main:continued-attachments",
-    );
-    expect(handoff?.send).toBe(true);
-    expect(handoff?.attachments).toHaveLength(1);
-    expect(handoff?.attachments[0]).toMatchObject({
-      mimeType: "image/png",
-      fileName: "shot.png",
-      dataUrl: "data:image/png;base64,AAAA",
-    });
-  });
-
-  it("continues an attachment-only draft instead of silently ignoring the send", async () => {
-    const request = vi.fn().mockResolvedValue({ sessionKey: "agent:main:attachment-only" });
-    const { key, pane, state } = createCatalogContinuationPane(request);
-    state.chatMessage = "";
-    state.chatAttachments = [
-      {
-        id: "att-2",
-        mimeType: "image/png",
-        fileName: "only.png",
-        dataUrl: "data:image/png;base64,BBBB",
-      },
-    ];
-
-    await pane.continueCatalogSession(key);
-
-    expect(request).toHaveBeenCalledWith("sessions.catalog.continue", {
-      ...key,
-      agentId: "main",
-    });
-    const handoff = consumePaneSessionHandoff(
-      pane.context,
-      pane.paneId,
-      "agent:main:attachment-only",
-    );
-    expect(handoff?.send).toBe(true);
-    expect(handoff?.attachments).toHaveLength(1);
-  });
-
-  it("still ignores a continuation with no draft and no attachments", async () => {
-    const request = vi.fn();
-    const { key, pane, state } = createCatalogContinuationPane(request);
-    state.chatMessage = "   ";
-    state.chatAttachments = [];
-
-    await pane.continueCatalogSession(key);
-
-    expect(request).not.toHaveBeenCalled();
-  });
-
-  it("does not stage or send a continuation rejected by its logical pane", async () => {
-    const request = vi.fn().mockResolvedValue({ sessionKey: "agent:main:rejected-continuation" });
-    const { key, pane, state } = createCatalogContinuationPane(request);
-    pane.onPaneSessionChange = vi.fn(() => false);
-
-    await pane.continueCatalogSession(key);
-
-    expect(
-      consumePaneSessionHandoff(pane.context, pane.paneId, "agent:main:rejected-continuation"),
-    ).toBeNull();
-    expect(state.handleSendChat).not.toHaveBeenCalled();
-    expect(state.chatSending).toBe(false);
-  });
-
-  it("does not send a stale catalog draft after the user switches conversations", async () => {
-    const continued = createDeferred<{ sessionKey: string }>();
-    const request = vi.fn(() => continued.promise);
-    const { key, pane, state } = createCatalogContinuationPane(request);
-
-    const pending = pane.continueCatalogSession(key);
-    state.sessionKey = "agent:main:different-conversation";
-    pane.sessionKey = state.sessionKey;
-    pane.catalogLoadGeneration += 1;
-    state.chatMessage = "Draft belonging to the selected conversation";
-    continued.resolve({ sessionKey: "agent:main:stale-continuation" });
-    await pending;
-
-    expect(pane.onPaneSessionChange).not.toHaveBeenCalled();
-    expect(state.handleChatDraftChange).not.toHaveBeenCalled();
-    expect(state.handleSendChat).not.toHaveBeenCalled();
-    expect(state.sessionKey).toBe("agent:main:different-conversation");
-    expect(state.chatMessage).toBe("Draft belonging to the selected conversation");
-    expect(state.chatSending).toBe(false);
-  });
-
-  it("does not apply a catalog continuation after the selected agent changes", async () => {
-    const continued = createDeferred<{ sessionKey: string }>();
-    const request = vi.fn(() => continued.promise);
-    const { key, pane, state } = createCatalogContinuationPane(request);
-    state.assistantAgentId = "main";
-
-    const pending = pane.continueCatalogSession(key);
-    state.assistantAgentId = "jarvis";
-    continued.resolve({ sessionKey: "agent:main:stale-owner" });
-    await pending;
-
-    expect(request).toHaveBeenCalledWith("sessions.catalog.continue", {
-      ...key,
-      agentId: "main",
-    });
-    expect(pane.onPaneSessionChange).not.toHaveBeenCalled();
-    expect(state.handleChatDraftChange).not.toHaveBeenCalled();
-    expect(state.handleSendChat).not.toHaveBeenCalled();
-    expect(state.chatSending).toBe(false);
-  });
-
-  it("does not send a stale catalog draft after reconnecting the same Gateway client", async () => {
-    const continued = createDeferred<{ sessionKey: string }>();
-    const request = vi.fn(() => continued.promise);
-    const { key, pane, state } = createCatalogContinuationPane(request);
-
-    const pending = pane.continueCatalogSession(key);
-    pane.connectionGeneration += 1;
-    state.connectionEpoch = pane.connectionGeneration;
-    state.chatMessage = "Draft from the reconnected conversation";
-    continued.resolve({ sessionKey: "agent:main:stale-continuation" });
-    await pending;
-
-    expect(pane.onPaneSessionChange).not.toHaveBeenCalled();
-    expect(state.handleChatDraftChange).not.toHaveBeenCalled();
-    expect(state.handleSendChat).not.toHaveBeenCalled();
-    expect(state.chatMessage).toBe("Draft from the reconnected conversation");
-    expect(state.chatSending).toBe(false);
-  });
-
-  it("does not apply an old catalog continuation after replacing the Gateway client", async () => {
-    const continued = createDeferred<{ sessionKey: string }>();
-    const request = vi.fn(() => continued.promise);
-    const { key, pane, sessions, state } = createCatalogContinuationPane(request);
-    const replacementClient = { request: vi.fn() } as unknown as GatewayBrowserClient;
-
-    const pending = pane.continueCatalogSession(key);
-    state.client = replacementClient;
-    pane.connectedClient = replacementClient;
-    pane.context = createSessionContext(replacementClient, sessions);
-    pane.connectionGeneration += 1;
-    state.connectionEpoch = pane.connectionGeneration;
-    state.chatMessage = "Draft from the replacement Gateway";
-    continued.resolve({ sessionKey: "agent:main:stale-continuation" });
-    await pending;
-
-    expect(pane.onPaneSessionChange).not.toHaveBeenCalled();
-    expect(state.handleChatDraftChange).not.toHaveBeenCalled();
-    expect(state.handleSendChat).not.toHaveBeenCalled();
-    expect(state.client).toBe(replacementClient);
-    expect(state.chatMessage).toBe("Draft from the replacement Gateway");
-    expect(state.chatSending).toBe(false);
-  });
-
-  it("does not clear a newer scoped send when a stale catalog continuation resolves", async () => {
-    const continued = createDeferred<{ sessionKey: string }>();
-    const request = vi.fn(() => continued.promise);
-    const { key, pane, state } = createCatalogContinuationPane(request);
-
-    const pending = pane.continueCatalogSession(key);
-    state.sessionKey = "agent:main:different-conversation";
-    pane.sessionKey = state.sessionKey;
-    pane.catalogLoadGeneration += 1;
-    state.chatSendingScopeKey = "newer-conversation-send";
-    state.chatSending = true;
-    continued.resolve({ sessionKey: "agent:main:stale-continuation" });
-    await pending;
-
-    expect(state.handleSendChat).not.toHaveBeenCalled();
-    expect(state.chatSendingScopeKey).toBe("newer-conversation-send");
-    expect(state.chatSending).toBe(true);
-  });
-
-  it("allows only the latest overlapping catalog continuation to adopt and send", async () => {
-    const first = createDeferred<{ sessionKey: string }>();
-    const second = createDeferred<{ sessionKey: string }>();
-    const request = vi
-      .fn()
-      .mockImplementationOnce(() => first.promise)
-      .mockImplementationOnce(() => second.promise);
-    const { key, pane, state } = createCatalogContinuationPane(request);
-
-    const staleContinuation = pane.continueCatalogSession(key);
-    state.chatMessage = "Only send the latest catalog draft";
-    const currentContinuation = pane.continueCatalogSession(key);
-    first.resolve({ sessionKey: "agent:main:stale-continuation" });
-    await staleContinuation;
-
-    expect(state.handleSendChat).not.toHaveBeenCalled();
-    expect(state.chatSending).toBe(true);
-
-    second.resolve({ sessionKey: "agent:main:latest-continuation" });
-    await currentContinuation;
-
-    expect(pane.onPaneSessionChange).toHaveBeenCalledOnce();
-    expect(
-      consumePaneSessionHandoff(pane.context, pane.paneId, "agent:main:latest-continuation"),
-    ).toEqual({
-      attachments: [],
-      draft: "Only send the latest catalog draft",
-      send: true,
-    });
-    expect(state.handleSendChat).not.toHaveBeenCalled();
-  });
-
-  it("does not display a rejected catalog continuation in a different conversation", async () => {
-    const continued = createDeferred<{ sessionKey: string }>();
-    const request = vi.fn(() => continued.promise);
-    const { key, pane, requestUpdate, state } = createCatalogContinuationPane(request);
-
-    const pending = pane.continueCatalogSession(key);
-    state.sessionKey = "agent:main:different-conversation";
-    pane.sessionKey = state.sessionKey;
-    pane.catalogLoadGeneration += 1;
-    state.lastError = "Current conversation error";
-    state.chatMessage = "Draft belonging to the selected conversation";
-    const updatesBeforeReject = requestUpdate.mock.calls.length;
-    continued.reject(new Error("Stale catalog continuation failed"));
-    await pending;
-
-    expect(state.lastError).toBe("Current conversation error");
-    expect(state.chatSending).toBe(false);
-    expect(state.handleSendChat).not.toHaveBeenCalled();
-    expect(requestUpdate).toHaveBeenCalledTimes(updatesBeforeReject + 1);
-  });
-
-  it("reports a catalog continuation failure in the original conversation", async () => {
-    const request = vi.fn().mockRejectedValue(new Error("Catalog continuation failed"));
-    const { key, pane, state } = createCatalogContinuationPane(request);
-
-    await pane.continueCatalogSession(key);
-
-    expect(state.lastError).toBe("Catalog continuation failed");
-    expect(state.chatSending).toBe(false);
-    expect(state.handleSendChat).not.toHaveBeenCalled();
   });
 });
